+++ src/main/java/com/nhnacademy/lastfrontproject/adaptor/AuthAdaptor.java
@@ -3,13 +3,14 @@ package com.nhnacademy.lastfrontproject.adaptor;
 import com.nhnacademy.lastfrontproject.dto.auth.LoginRequest;
 import com.nhnacademy.lastfrontproject.dto.auth.RegisterRequest;
 import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-@FeignClient(name = "auth", url = "${external.auth-service.url}")
+@FeignClient(name = "auth", url = "${feign.client.gateway-service.url}")
 public interface AuthAdaptor {
     @PostMapping("/auth/register")
-    void register(@RequestBody RegisterRequest registerRequest);
+    ResponseEntity<String> register(@RequestBody RegisterRequest registerRequest);
 
     @PostMapping("/auth/login")
     void login(@RequestBody LoginRequest loginRequest);
