--- src/main/java/com/nhnacademy/lastfrontproject/controller/SignInController.java
@@ -1,12 +1,0 @@
-package com.nhnacademy.lastfrontproject.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class SignInController {
-    @GetMapping("/sign-in")
-    public String signIn() {
-        return "pages-sign-in";
-    }
-}
