--- src/main/java/com/nhnacademy/lastfrontproject/controller/SignUpController.java
@@ -1,12 +1,0 @@
-package com.nhnacademy.lastfrontproject.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class SignUpController {
-    @GetMapping("/sign-up")
-    public String signUp() {
-        return "pages-sign-up";
-    }
-}
